+++ tests/ADN.Helpers.Tests/Models/Data/Base64HelperTest.cs
@@ -13,18 +13,18 @@ namespace ADN.Helpers.Tests.Data
         [Theory]
         [InlineData("0123456789", "MDEyMzQ1Njc4OQ==")]
         [InlineData("This is a test", "VGhpcyBpcyBhIHRlc3Q=")]
-        public void Encode(string value, string expected)
+        public void Base64Encode(string value, string expected)
         {
-            var result = Base64Helper.Encode(value);
+            var result = value.Base64Encode();
             Assert.Equal(expected, result);
         }
 
         [Theory]
         [InlineData("MDEyMzQ1Njc4OQ==", "0123456789")]
         [InlineData("VGhpcyBpcyBhIHRlc3Q=", "This is a test")]
-        public void Decode(string value, string expected)
+        public void Base64Decode(string value, string expected)
         {
-            var result = Base64Helper.Decode(value);
+            var result = value.Base64Decode();
             Assert.Equal(expected, result);
         }
     }